added form shake on invalid submit
Omitted resetting the forms, in order to increase accessibility to the user (i.e. show them WHAT input was invalid). Instead, an error message asking for valid credentials is sent to the login form.

Additionally, specified the signup form's error section in the updatePasswordValidation function - as the login form shouldn't display these messages.